Split routes tests into separate files
